Place Minikits first/last depending on accessibility to improve fill
For minimal players, place Minikits first. This helps prevent fill from
dumping non-macguffins into unreachable locations and reducing the
number of reachable locations to fewer than the number of items
remaining to be placed.

For non-minimal players, place Minikits last. The helps prevent fill
from filling most/all reachable locations with the macguffins and having
to swap some out that are only required for the goal.